+++ tket/tests/Circuit/test_Multiplexor.cpp
@@ -203,20 +203,20 @@ SCENARIO("MultiplexedRotationBox decomposition", "[boxes]") {
   }
 }
 
-SCENARIO("UniformQControlU2Box decomposition", "[boxes]") {
-  GIVEN("UniformQControlU2Box with zero control") {
+SCENARIO("MultiplexedU2Box decomposition", "[boxes]") {
+  GIVEN("MultiplexedU2Box with zero control") {
     ctrl_op_map_t op_map = {{{}, get_op_ptr(OpType::H)}};
-    UniformQControlU2Box uqcu2_box(op_map);
-    std::shared_ptr<Circuit> c = uqcu2_box.to_circuit();
+    MultiplexedU2Box multiplexor(op_map);
+    std::shared_ptr<Circuit> c = multiplexor.to_circuit();
     std::vector<Command> cmds = c->get_commands();
     REQUIRE(cmds.size() == 1);
     REQUIRE(check_multiplexor(op_map, *c));
   }
-  GIVEN("UniformQControlU2Box with one control") {
+  GIVEN("MultiplexedU2Box with one control") {
     ctrl_op_map_t op_map = {
         {{1}, get_op_ptr(OpType::H)}, {{0}, get_op_ptr(OpType::X)}};
-    UniformQControlU2Box uqcu2_box(op_map);
-    std::shared_ptr<Circuit> c = uqcu2_box.to_circuit();
+    MultiplexedU2Box multiplexor(op_map);
+    std::shared_ptr<Circuit> c = multiplexor.to_circuit();
     std::vector<Command> cmds = c->get_commands();
     REQUIRE(cmds.size() == 3 + 1);
     for (auto cmd : cmds) {
@@ -227,7 +227,7 @@ SCENARIO("UniformQControlU2Box decomposition", "[boxes]") {
     }
     REQUIRE(check_multiplexor(op_map, *c));
   }
-  GIVEN("UniformQControlU2Box with two controls") {
+  GIVEN("MultiplexedU2Box with two controls") {
     Circuit c0(1);
     c0.add_op<unsigned>(OpType::TK1, {0.2374, 1.0353, 0.5372}, {0});
     Eigen::Matrix2cd m = tket_sim::get_unitary(c0);
@@ -237,8 +237,8 @@ SCENARIO("UniformQControlU2Box decomposition", "[boxes]") {
         {{1, 1}, mbox_op},
         {{0, 1}, get_op_ptr(OpType::X)},
         {{1, 0}, get_op_ptr(OpType::TK1, std::vector<Expr>{0.3, 1.8, 3.4})}};
-    UniformQControlU2Box uqcu2_box(op_map);
-    std::shared_ptr<Circuit> c = uqcu2_box.to_circuit();
+    MultiplexedU2Box multiplexor(op_map);
+    std::shared_ptr<Circuit> c = multiplexor.to_circuit();
     std::vector<Command> cmds = c->get_commands();
     REQUIRE(cmds.size() == 7 + 2);
     for (auto cmd : cmds) {
@@ -249,7 +249,7 @@ SCENARIO("UniformQControlU2Box decomposition", "[boxes]") {
     }
     REQUIRE(check_multiplexor(op_map, *c));
   }
-  GIVEN("UniformQControlU2Box with 5 controls") {
+  GIVEN("MultiplexedU2Box with 5 controls") {
     ctrl_op_map_t op_map = {
         {{1, 1, 0, 1, 0}, get_op_ptr(OpType::H)},
         {{0, 1, 1, 0, 0}, get_op_ptr(OpType::X)},
@@ -258,8 +258,8 @@ SCENARIO("UniformQControlU2Box decomposition", "[boxes]") {
         {{1, 0, 1, 0, 1},
          get_op_ptr(OpType::TK1, std::vector<Expr>{0.3, 1.8, 3.4})},
     };
-    UniformQControlU2Box uqcu2_box(op_map);
-    std::shared_ptr<Circuit> c = uqcu2_box.to_circuit();
+    MultiplexedU2Box multiplexor(op_map);
+    std::shared_ptr<Circuit> c = multiplexor.to_circuit();
     std::vector<Command> cmds = c->get_commands();
     REQUIRE(cmds.size() == 63 + 5);
     for (auto cmd : cmds) {
@@ -270,11 +270,11 @@ SCENARIO("UniformQControlU2Box decomposition", "[boxes]") {
     }
     REQUIRE(check_multiplexor(op_map, *c));
   }
-  GIVEN("UniformQControlU2Box without the final diagonal") {
+  GIVEN("MultiplexedU2Box without the final diagonal") {
     ctrl_op_map_t op_map = {
         {{1}, get_op_ptr(OpType::H)}, {{0}, get_op_ptr(OpType::X)}};
-    UniformQControlU2Box uqcu2_box(op_map, false);
-    std::shared_ptr<Circuit> c = uqcu2_box.to_circuit();
+    MultiplexedU2Box multiplexor(op_map, false);
+    std::shared_ptr<Circuit> c = multiplexor.to_circuit();
     std::vector<Command> cmds = c->get_commands();
     REQUIRE(cmds.size() == 3);
   }
@@ -334,28 +334,28 @@ SCENARIO("Exception handling", "[boxes]") {
         MultiplexedRotationBox(op_map), BadOpType,
         MessageContains("Ops must be either Rx, Ry, or Rz"));
   }
-  GIVEN("UniformQControlU2Box unsupported gate") {
+  GIVEN("MultiplexedU2Box unsupported gate") {
     ctrl_op_map_t op_map = {
         {{0, 1}, get_op_ptr(OpType::H)}, {{1, 0}, get_op_ptr(OpType::CX)}};
     REQUIRE_THROWS_MATCHES(
-        UniformQControlU2Box(op_map), BadOpType,
+        MultiplexedU2Box(op_map), BadOpType,
         MessageContains(
             "Ops must be single-qubit unitary gate types or Unitary1qBox"));
   }
-  GIVEN("Decompose symbolic UniformQControlU2Box") {
+  GIVEN("Decompose symbolic MultiplexedU2Box") {
     Sym a = SymTable::fresh_symbol("a");
     Expr expr_a(a);
     ctrl_op_map_t op_map = {{{1}, get_op_ptr(OpType::Ry, expr_a)}};
-    UniformQControlU2Box u2box(op_map);
+    MultiplexedU2Box u2box(op_map);
     REQUIRE_THROWS_MATCHES(
         u2box.to_circuit(), Unsupported,
-        MessageContains("Can't decompose symbolic UniformQControlU2Box"));
+        MessageContains("Can't decompose symbolic MultiplexedU2Box"));
   }
 }
 
 TEMPLATE_TEST_CASE(
     "Auxiliary methods", "[boxes]", MultiplexorBox, MultiplexedRotationBox,
-    UniformQControlU2Box) {
+    MultiplexedU2Box) {
   GIVEN("symbol_substitution") {
     Sym a = SymTable::fresh_symbol("a");
     Expr expr_a(a);